`load_data` and plotting and most of `fit`
`load_data` now takes a list of column numbers for fitting with multiple
features.

Added most of the code for the `fit` along with `predict` and `net_input`.

Wrote in all the code from PDF for plotting the result plots with the
fits, but seems to be misaligned on x axis.